remove apis/stack.py
