+++ docs/openapi_generator/generate.py
@@ -31,7 +31,34 @@ from .pyopenapi.utility import Specification
 
 schema_utils.json_schema_type = json_schema_type
 
-from llama_stack.apis.stack import LlamaStack
+from llama_models.llama3.api.datatypes import *  # noqa: F403
+from llama_stack.apis.agents import *  # noqa: F403
+from llama_stack.apis.dataset import *  # noqa: F403
+from llama_stack.apis.evals import *  # noqa: F403
+from llama_stack.apis.inference import *  # noqa: F403
+from llama_stack.apis.batch_inference import *  # noqa: F403
+from llama_stack.apis.memory import *  # noqa: F403
+from llama_stack.apis.telemetry import *  # noqa: F403
+from llama_stack.apis.post_training import *  # noqa: F403
+from llama_stack.apis.reward_scoring import *  # noqa: F403
+from llama_stack.apis.synthetic_data_generation import *  # noqa: F403
+from llama_stack.apis.safety import *  # noqa: F403
+
+
+class LlamaStack(
+    Inference,
+    BatchInference,
+    Agents,
+    RewardScoring,
+    Safety,
+    SyntheticDataGeneration,
+    Datasets,
+    Telemetry,
+    PostTraining,
+    Memory,
+    Evaluations,
+):
+    pass
 
 
 # TODO: this should be fixed in the generator itself so it reads appropriate annotations
